Add Middleware and Comment Route

+++ routes/api.php
@@ -1,23 +1,29 @@
 <?php
 
 use App\Http\Controllers\AuthenticationController;
+use App\Http\Controllers\CommentController;
 use App\Http\Controllers\MenuController;
 use App\Http\Controllers\OrderController;
+use App\Http\Middleware\AdminAccess;
+use App\Http\Middleware\UserComment;
 use Illuminate\Support\Facades\Route;
 
 Route::middleware(['auth:sanctum'])->group(function () {
     Route::get('/logout', [AuthenticationController::class, 'logout']);
     Route::get('/me', [AuthenticationController::class, 'me']);
 
-    Route::post('/menus', [MenuController::class, 'store']);
-    Route::patch('/menus/{id}', [MenuController::class, 'update']);
-    Route::delete('/menus/{id}', [MenuController::class, 'destroy']);
+    Route::post('/menus', [MenuController::class, 'store'])->middleware(AdminAccess::class);
+    Route::patch('/menus/{id}', [MenuController::class, 'update'])->middleware(AdminAccess::class);
+    Route::delete('/menus/{id}', [MenuController::class, 'destroy'])->middleware(AdminAccess::class);
 
     
     Route::get('/orders', [OrderController::class, 'index']);
     Route::post('/orders', [OrderController::class, 'store']);
     Route::patch('/orders/{id}', [OrderController::class, 'update']);
     Route::delete('/orders/{id}', [OrderController::class, 'destroy']);
+
+    Route::post('/comments', [CommentController::class, 'store']);
+    Route::delete('/comments/{id}', [CommentController::class, 'destroy'])->middleware(UserComment::class);
 });
 
 Route::get('/menus', [MenuController::class, 'index']);
